update: added warnings and functionality
There was a change before that allowed this library to use more of `aLight` settings when using the `light` property when creating a light. It however never actually assigned those settings to the creation of the light, but rather allowed for them to be passed without error. This update adds that functionality.

Fixed a few broken references.
Created a new object for a variable rather than relying on it using the passed in reference. (this was causing issues with another variable, and was hard to track down) So it's proactive to do the same here.
Fixed a few variable type declarations to the proper type.