update: removed some code
Removed a check that would only run due to an engine bug.  The engine bug called `onNew` for elements that were added via `addInterfaceElement` even if they already existed prior. This would cause the code above to run twice causing all types of errors. So the code returned if it was already called prior, the engine bug was fixed causing this code to be obsolete now.